[dsa] reimplement Vector internal storage
The Dynamic_Array was changed to default initialise all of it's values
in order to be easier to work with allocators. This caused the unused
values of the Vector to be default initialised, which is something we do
not want. Since the Dynamic_Array and Vector storage has different
requirements a custom implementation for the Vector was created.