[dsa] Introduce Monitor class
This class allows code to peek into how containers are managing memory.
We need this functionality from both the visualisation side and the test
side of the code. The former needs to visualise the memory of each
container while the former tests that the allocation and object life
time operations are not miused.