[dsa] Make Memory_Monitor handler calls static
The Memory_Monitor previous had a static instance inside of the class.
Since it takes a type template parameter this resulted in multiple
static instances being accessed for different types. This creates a
problem if an allocator is created with a different type. An example is
a list which creates an allocator for its internal Node type. Having the
event calls static the Handler can decide what to do, if it creates a
static instance then it will be shared by all types.