[dsa] Make Element_Monitor inherit the wrapped value
The Element_Monitor should behave as the wrapped value. For containers
that allocate a custom class (such as a node in a list) an is-a
relationship allows the container to directly access the fields in that
type. Otherwise we need to pollute the code with casts to access the
expected fields.