[test] allocation_verifier: Support monitoring of structures
When monitoring a structure we receive a construct call for every field
within the structure. This make the previous implementation think that
the same memory was being constructed multiple times without it being
destroyed. We now dynamically build a representation of the memory and
figure out when the subfields are being constructed.