[dsa] Send element value to memory_representation when it is changed
Since we can process the events separately from the time at which they
are dispatched we do not know if pointer accesses are safe. Since we
need to know the value for certain operations we send a copy of it when
the event is dispatched, this makes sure we do not need to access it
later.