[visual] add mechanism to dynamically disable/ enable events
There are places where we run operations on the data structures in order
to read its state. These operations emit a lot of events which the user
interface then renders, the problem is that since the user did not
explicitly start these events it is confusing as to why they are being
rendered. With this functionality we can disable the events for the
background operations and enable them for the operations initiated by
the user.